web/requirements.txt
pyyaml
python-decouple==3
